moves files inside 'schedule' to match the rest of the project
